Locks.jl

Promise

API	Summary

Promise	Create a promise of type T; i.e., a memory location that holds a value of type T that can be set once and retrieved asynchronously.

try_race_fetch	Try to retrieve a value if it is already set. Return Ok(value) on success and Err(NotSetError()) on failure.

try_race_fetch_or!	Fetch an existing value or set a computed value (computed = thunk()). The thunk is called at most once for each instance of promise.

race_fetch_or!	Fetch an existing value or set value = thunk(). The thunk is called at most once for each instance of promise.

try_race_put!	Try to set a value in the promise.

Locks.Promise — Type

Promise{T}()
Promise()

Create a promise of type T; i.e., a memory location that holds a value of type T that can be set once and retrieved asynchronously.

Extended help

For an explanation on the concept, see: Futures and promises - Wikipedia.

Examples

julia> using Locks

julia> p = Promise{Int}();

julia> try_race_fetch(p)
Try.Err: NotSetError()

julia> put!(p, 123);

julia> fetch(p)
123

julia> try_race_put!(p, 456)
Try.Err: OccupiedError{Int64}(123)

Memory ordering

An event that retrieves or waits for a value from a promise establishes a happened-before edge from the event that have set the value to the promise. Invocations of the API that includes an event that sets a value to a promise include:

put!(promise, value) that does not throw.

try_race_put!(promise, value) that returns an Ok result.

try_race_fetch_or!(thunk, promise) that calls thunk (and hance returns an Err).

race_fetch_or!(thunk, promise) that calls thunk.

(The invocation thunk() is sequenced-before the event that sets the value.)

Invocations of the API that includes an event that retrieves or waits for a value from a promise include:

fetch(promise)

wait(promise)

try_race_fetch(promise) that returns an Ok result.

try_race_fetch_or!(thunk, promise) that does not call thunk (and hance returns an Ok)

race_fetch_or!(thunk, promise) that does not call thunk

Supported operations

A promise::Promise{T} supports the following operations:

put!(promise, value): Set a value; throws an error if some value has already been set.

try_race_put!(promise, value): Try to set a value.

fetch(promise): Obtain a value that the promise holds. Wait if necessary.

wait(promise): Wait for a value to be set.

try_race_fetch(promise): Try to obtain an existing value in the promise.

try_race_fetch_or!(thunk, promise): Try to get a value or set a value generated by thunk.

race_fetch_or!(thunk, promise): Similar to try_race_fetch_or! but the returned value does not indicate if thunk is called or not.

Locks.try_race_fetch — Function

try_race_fetch(promiselike) -> Ok(value::T) or Err(NotSetError())

Try to retrieve a value if it is already set. Return Ok(value) on success and Err(NotSetError()) on failure.

try_race_fetch can be called on a Promise or a tasklet (@tasklet).

Extended help

Examples

julia> using Locks

julia> p = Promise{Int}();

julia> try_race_fetch(p)
Try.Err: NotSetError()

julia> put!(p, 123);

julia> try_race_fetch(p)
Try.Ok: 123

Locks.try_race_fetch_or! — Function

try_race_fetch_or!(thunk, promise::Promise{T}) -> Ok(existing::T) or Err(computed::T)

Fetch an existing value or set a computed value (computed = thunk()). The thunk is called at most once for each instance of promise.

Extended help

Examples

julia> using Locks

julia> p = Promise{Int}();

julia> try_race_fetch_or!(p) do
           123 + 456
       end
Try.Err: 579

julia> try_race_fetch_or!(p) do
           42
       end
Try.Ok: 579

Locks.race_fetch_or! — Function

race_fetch_or!(thunk, promise::Promise{T}) -> value::T

Fetch an existing value or set value = thunk(). The thunk is called at most once for each instance of promise.

This is similar to try_race_fetch_or! but the caller cannot tell if thunk is called or not by the return type.

Extended help

Examples

julia> using Locks

julia> p = Promise{Int}();

julia> race_fetch_or!(p) do
           println("called")
           123 + 456
       end
called
579

julia> race_fetch_or!(p) do
           println("called")
           42
       end
579

Locks.try_race_put! — Function

try_race_put!(promise::Promise{T}, value) -> Ok(value′::T) or Err(OccupiedError(existing::T))

Try to set a value in the promise.

Since the value is converted to T first, the returned value′ may not be identical to the input value.

Extended help

Examples

julia> using Locks

julia> p = Promise{Int}();

julia> try_race_put!(p, 123)
Try.Ok: 123

julia> try_race_put!(p, 456)
Try.Err: OccupiedError{Int64}(123)

Promise-like interfaces

API	Summary

@tasklet	Create an object that is a memoized version of () -> code but also acts like a Promise that is not settable.

@once	Execute code at most once within the lifetime of the Julia process. It evaluates to the result of the first evaluation whenever the same location of the code is evaluated later.

Locks.@tasklet — Macro

@tasklet code

Create an object that is a memoized version of () -> code but also acts like a Promise that is not settable.

A t = @tasklet code supports: t(), fetch(t), wait(t), and try_race_fetch.

Extended help

Examples

julia> using Locks

julia> t = @tasklet begin
           println("called")
           123
       end;

julia> try_race_fetch(t)
Try.Err: NotSetError()

julia> t()
called
123

julia> t()
123

julia> fetch(t)
123

julia> wait(t);

Memory ordering

An event that retrieves or waits for a value from a tasklet t establishes a happened-before edge from the events in code. Invocations of the API that includes an event that retrieves or waits for a value from a promise include:

fetch(t)

wait(t)

try_race_fetch(t) that returns an Ok result.

Supported operations

A tasklet t = @tasklet code supports the following operations:

t(): Evaluate code if it hasn't been evaluated. Otherwise, equivalent to fetch.

fetch(t): Wait for other tasks to invoke t() and then return the result.

wait(t): Wait for other tasks to invoke t().

try_race_fetch: Try to retrieve the result of t() if it is already called.

Locks.@once — Macro

@once code

Execute code at most once within the lifetime of the Julia process. It evaluates to the result of the first evaluation whenever the same location of the code is evaluated later.

Note that code does not have access to the local variables.

Memory ordering

Events that follow the run-time (as opposed to macro expansion-time) evaluation of @once code in the same Julia task have happened-before edge from the events in code.

Examples

julia> using Locks

julia> f() = @once Ref(123);

julia> f() isa Ref
true

julia> f()[]
123

Ref(123) is evaluated once and the identical Ref object is reused whenever f is called in this Julia process:

julia> f() === f()
true

Locks

API	Summary

ReentrantCLHLock	A reentrant CLH "spinnable" lock that provides first-come-first-served fairness. Keyword argument nspins::Integer can be passed to acquire to specify a number of spins tried before falling back to "cooperative" waiting in the Julia scheduler.

NonreentrantCLHLock	A (non-reentrant) CLH "spinnable" lock that provides first-come-first-served fairness. Keyword argument nspins::Integer can be passed to acquire to specify a number of spins tried before falling back to "cooperative" waiting in the Julia scheduler.

ReentrantBackoffSpinLock	A reentrant exponential backoff spin lock.

NonreentrantBackoffSpinLock	A non-reentrant exponential backoff spin lock.

TaskObliviousLock	A lock that can be released in a task that did not acquire the lock. It does not support reentrancy.

read_write_lock	Return the read handle rlock and the write handle wlock of a read-write lock.

acquire	Acquire a lock. It is equivalent to Base.lock(lock) but it may support additional keyword arguments.

release	Release a lock. It is equivalent to Base.unlock(lock).

try_race_acquire	Try to acquire lock and return Ok(nothing) on success. Return an Err wrapping a value explaining a reason of failure.

race_acquire	Try to acquire lock and return true on success and false on failure.

acquire_then	Execute a thunk f in a critical section protected by lock and return the value y returned from f. Keyword arguments are passed to acquire.

Locks.ReentrantCLHLock — Type

ReentrantCLHLock

A reentrant CLH "spinnable" lock that provides first-come-first-served fairness. Keyword argument nspins::Integer can be passed to acquire to specify a number of spins tried before falling back to "cooperative" waiting in the Julia scheduler.

Extended help

ReentrantCLHLock implements the spin lock by Craig (1993) and Magnussen, Landin, and Hagersten (1994) with a fallback to "cooperatively" wait in the scheduler instead of spinning (hence "spinnable"). See NonreentrantCLHLock that provides a non-reentrant version.

Memory ordering

A release invocation on a lock establishes happened-before edges to subsequent invocations of acquire and try_race_acquire that returns an Ok on the same lock.

Supported operations

acquire(lock::ReentrantCLHLock; [nspins::Integer]) (lock)

try_race_acquire(lock::ReentrantCLHLock) (trylock): Not very efficient but lock-free. Fail with AcquiredByWriterError.

release(lock::ReentrantCLHLock) (unlock)

Locks.NonreentrantCLHLock — Type

NonreentrantCLHLock

A (non-reentrant) CLH "spinnable" lock that provides first-come-first-served fairness. Keyword argument nspins::Integer can be passed to acquire to specify a number of spins tried before falling back to "cooperative" waiting in the Julia scheduler.

Extended help

NonreentrantCLHLock implements the spin lock by Craig (1993) and Magnussen, Landin, and Hagersten (1994) with a fallback to "cooperatively" wait in the scheduler instead of spinning (hence "spinnable"). See ReentrantCLHLock that provides a reentrant version.

Memory ordering

NonreentrantCLHLock has the same semantics as ReentrantCLHLock provided that each hand-off of the lock between tasks (if any) establishes a happened-before edge.

Supported operations

acquire(lock::NonreentrantCLHLock; [nspins::Integer]) (lock)

try_race_acquire(lock::NonreentrantCLHLock) (trylock): Not very efficient but lock-free. Fail with AcquiredByWriterError.

release(lock::NonreentrantCLHLock) (unlock)

Locks.ReentrantBackoffSpinLock — Type

ReentrantBackoffSpinLock

A reentrant exponential backoff spin lock.

See also NonreentrantBackoffSpinLock that provides a non-reentrant version.

Extended help

Memory ordering

A release invocation on a lock establishes happened-before edges to subsequent invocations of acquire and try_race_acquire that returns an Ok on the same lock.

Supported operations

ReentrantBackoffSpinLock(; [mindelay], [maxdelay]) -> lock: Create a lock. mindelay (default: 1) specifies the number of spinloop called in the initial backoff. mindelay (default: 1000) specifies the maximum backoff.

acquire(lock::ReentrantBackoffSpinLock; [mindelay], [maxdelay]) (lock): Acquire the lock. Keyword arguments mindelay and maxdelay can be passed to override the values specified by the constructor.

try_race_acquire(lock::ReentrantBackoffSpinLock) (trylock):

release(lock::ReentrantBackoffSpinLock) (unlock)

Locks.NonreentrantBackoffSpinLock — Type

NonreentrantBackoffSpinLock

A non-reentrant exponential backoff spin lock.

See also ReentrantBackoffSpinLock that provides a reentrant version.

Extended help

NonreentrantBackoffSpinLock performs better than Base.Threads.SpinLock with high contention. NonreentrantCLHLock is better than NonreentrantBackoffSpinLock with high contention with many worker threads (20 to 80; it depends on the machine).

Since NonreentrantBackoffSpinLock does not have a fallback "cooperative" waiting mechanism, NonreentrantCLHLock is in general recommended.

Memory ordering

NonreentrantBackoffSpinLock has the same semantics as ReentrantBackoffSpinLock provided that each hand-off of the lock between tasks (if any) establishes a happened-before edge.

Supported operations

NonreentrantBackoffSpinLock(; [mindelay], [maxdelay]) -> lock: Create a lock. mindelay (default: 1) specifies the number of spinloop called in the initial backoff. mindelay (default: 1000) specifies the maximum backoff.

acquire(lock::NonreentrantBackoffSpinLock; [mindelay], [maxdelay]) (lock): Acquire the lock. Keyword arguments mindelay and maxdelay can be passed to override the values specified by the constructor.

try_race_acquire(lock::NonreentrantBackoffSpinLock) (trylock):

release(lock::NonreentrantBackoffSpinLock) (unlock)

Locks.TaskObliviousLock — Type

TaskObliviousLock

A lock that can be released in a task that did not acquire the lock. It does not support reentrancy.

Extended help

Examples

julia> using Locks

julia> lock = TaskObliviousLock();

julia> acquire(lock);

julia> wait(Threads.@spawn release(lock));  # completes

Supported operations

acquire(lock::TaskObliviousLock) (lock)

release(lock::TaskObliviousLock) (unlock)

Implementation detail

TaskObliviousLock is an alias to unspecified implementation of lock. Currently, it is:

julia> TaskObliviousLock === NonreentrantCLHLock
true

Locks.read_write_lock — Function

read_write_lock() -> (rlock, wlock)

Return the read handle rlock and the write handle wlock of a read-write lock.

Extended help

Supported operations:

acquire(rlock) (lock)

try_race_acquire(rlock; [nspins], [ntries]) (trylock): Not very efficient but lock-free. Fail with TooManyTries.

release(rlock) (unlock)

acquire(wlock) (lock)

try_race_acquire(wlock) (trylock): Fail with NotAcquirableError.

release(wlock) (unlock)

Locks.acquire — Function

acquire(lock)

Acquire a lock. It is equivalent to Base.lock(lock) but it may support additional keyword arguments.

See also release and try_race_acquire.

Extended help

Examples

julia> using Locks

julia> lock = ReentrantCLHLock();

julia> acquire(lock);

julia> release(lock);

On naming

Locks.jl uses acquire/release instead of lock/unlock so that:

Variable lock can be used.

Make it clear that Locks.try_race_acquire(lock) -> result::Union{Ok,Err} and Base.trylock(lock) -> locked::Bool have different return types. In particular, try_race_acquire can report the reason why certain attempt have failed.

Locks.release — Function

release(lock)

Release a lock. It is equivalent to Base.unlock(lock).

See also acquire and try_race_acquire.

Locks.try_race_acquire — Function

try_race_acquire(lock) -> Ok(nothing) or Err(reason)

Try to acquire lock and return Ok(nothing) on success. Return an Err wrapping a value explaining a reason of failure.

See the documentation of typeof(lock) for possible error types.

Examples

julia> using Locks

julia> lock = NonreentrantCLHLock();

julia> try_race_acquire(lock)
Try.Ok: nothing

julia> try_race_acquire(lock)
Try.Err: TooManyTries(0, 0)

Locks.race_acquire — Function

race_acquire(lock) -> isacquired::Bool

Try to acquire lock and return true on success and false on failure.

See also try_race_acquire.

Examples

julia> using Locks

julia> lock = NonreentrantCLHLock();

julia> race_acquire(lock)
true

julia> race_acquire(lock)
false

Locks.acquire_then — Function

acquire_then(f, lock; acquire_options...) -> y

Execute a thunk f in a critical section protected by lock and return the value y returned from f. Keyword arguments are passed to acquire.

Extended help

Examples

julia> using Locks

julia> lock = ReentrantCLHLock();

julia> acquire_then(lock; nspins = 10) do
           123 + 456
       end
579

Guards

API	Summary

Guard	Guard mutable data. Use guarding to obtain exclusive access to data.

ReadWriteGuard	Guard mutable data. Use guarding and guarding_read to obtain exclusive ("write") and shared ("read") access to data.

guarding	Apply f! to the data wrapped in guard while obtaining exclusive access.

guarding_read	Apply f to the data wrapped in guard while obtaining shared access.

Locks.Guard — Type

Guard(data)

Guard mutable data. Use guarding to obtain exclusive access to data.

Extended help

Examples

julia> using Locks

julia> guard = Guard(Ref(0));

julia> guarding(guard) do ref
           ref[] += 1
       end
1

Locks.ReadWriteGuard — Type

ReadWriteGuard(data)

Guard mutable data. Use guarding and guarding_read to obtain exclusive ("write") and shared ("read") access to data.

Extended help

Examples

julia> using Locks

julia> guard = ReadWriteGuard(Ref(0));

julia> guarding(guard) do ref
           ref[] += 1
       end
1

julia> guarding_read(guard) do ref
           ref[]  # must not mutate anything
       end
1

Locks.guarding — Function

guarding(f!, guard)

Apply f! to the data wrapped in guard while obtaining exclusive access.

See: Guard, ReadWriteGuard, guarding_read

Extended help

Examples

julia> using Locks

julia> guard = Guard(Ref(0));

julia> guarding(guard) do ref
           ref[] += 1
       end
1

Locks.guarding_read — Function

guarding_read(f, guard)

Apply f to the data wrapped in guard while obtaining shared access.

See: Guard, ReadWriteGuard, guarding

Extended help

Examples

julia> using Locks

julia> guard = ReadWriteGuard(Ref(0));

julia> guarding_read(guard) do ref
           ref[]  # must not mutate anything
       end
0

Low-level interfaces

API	Summary

ThreadLocalStorage	Create a thread-local storage of type T created by factory().

spinloop	A hint to the compiler, runtime, and hardware that spinloop() is in the middle of a spin loop. Call this in a spin loop that requires some other worker threads to make forward progress in order for the current thread to make forward progress.

Locks.ThreadLocalStorage — Type

ThreadLocalStorage{T}(factory)
ThreadLocalStorage(factory)

Create a thread-local storage of type T created by factory().

An instance tls of ThreadLocalStorage support the operation x = tls[] for obtaining an object x of value T.

Warning

Using this API is extremely tricky. Arguably, it is not even well-defined when and how it can be used.

Theoretically, it is safe to use this API if the programmer can ensure that, once a value x = tls[] is obtained, the code does not hit any yield points until there is no more access to x. However, it is not possible to know if a certain operation is yield-free in general.

Thus, this API currently exists primarily for helping migration of code written using nthreads and threadid in an ad-hoc manner.

An object of type T is allocated for each worker thread of the Julia runtime. If T is not given, T = typeof(factory()) is used (i.e., factory is assumed to be type-stable).

Extended help

Examples

julia> using Locks

julia> tls = ThreadLocalStorage(Ref{Int});

julia> tls[] isa Ref{Int}
true

julia> tls[][] = 123;

julia> tls[][]
123

Locks.spinloop — Function

spinloop()

A hint to the compiler, runtime, and hardware that spinloop() is in the middle of a spin loop. Call this in a spin loop that requires some other worker threads to make forward progress in order for the current thread to make forward progress.

Warning

Observe that the above sentence specifically mentions worker threads and not Tasks. A Julia programmer should always be alarmed whenever an API talks about threads instead of Tasks. Prefer higher-level APIs such as channels and condition variables.

A proper use of spinloop requires extra cares such as a fallback that waits in the Julia scheduler and/or a mechanism that enables the spin loop code path given that other threads exist and a task that can break the spin loop is running or will be scheduled eventually.

Implementation detail

It calls GC.safepoint() and jl_cpu_pause.