Locks.jl

API	Summary

NonreentrantBackoffSpinLock	A non-reentrant exponential backoff spin lock.

NonreentrantCLHLock	A (non-reentrant) CLH "spinnable" lock that provides first-come-first-served fairness. Keyword argument nspins::Integer can be passed to acquire to specify a number of spins tried before falling back to "cooperative" waiting in the Julia scheduler.

ReentrantBackoffSpinLock	A reentrant exponential backoff spin lock.

ReentrantCLHLock	A reentrant CLH "spinnable" lock that provides first-come-first-served fairness. Keyword argument nspins::Integer can be passed to acquire to specify a number of spins tried before falling back to "cooperative" waiting in the Julia scheduler.

TaskObliviousLock	A lock that can be released in a task that did not acquire the lock. It does not support reentrancy.

acquire	Acquire a lock. It is equivalent to Base.lock(lock) but it may support additional keyword arguments.

acquire_then	Execute a thunk f in a critical section protected by lock and return the value y returned from f. Keyword arguments are passed to acquire.

race_acquire	Try to acquire lock and return true on success and false on failure.

release	Release a lock. It is equivalent to Base.unlock(lock).

try_race_acquire	Try to acquire lock and return Ok(nothing) on success. Return an Err wrapping a value explaining a reason of failure.

Locks.ReentrantCLHLock — Type

ReentrantCLHLock

A reentrant CLH "spinnable" lock that provides first-come-first-served fairness. Keyword argument nspins::Integer can be passed to acquire to specify a number of spins tried before falling back to "cooperative" waiting in the Julia scheduler.

Extended help

ReentrantCLHLock implements the spin lock by Craig (1993) and Magnussen, Landin, and Hagersten (1994) with a fallback to "cooperatively" wait in the scheduler instead of spinning (hence "spinnable"). See NonreentrantCLHLock that provides a non-reentrant version.

Memory ordering

A release invocation on a lock establishes happened-before edges to subsequent invocations of acquire and try_race_acquire that returns an Ok on the same lock.

Supported operations

acquire(lock::ReentrantCLHLock; [nspins::Integer]) (lock)

try_race_acquire(lock::ReentrantCLHLock) (trylock): Not very efficient but lock-free. Fail with AcquiredByWriterError.

release(lock::ReentrantCLHLock) (unlock)

Locks.NonreentrantCLHLock — Type

NonreentrantCLHLock

A (non-reentrant) CLH "spinnable" lock that provides first-come-first-served fairness. Keyword argument nspins::Integer can be passed to acquire to specify a number of spins tried before falling back to "cooperative" waiting in the Julia scheduler.

Extended help

NonreentrantCLHLock implements the spin lock by Craig (1993) and Magnussen, Landin, and Hagersten (1994) with a fallback to "cooperatively" wait in the scheduler instead of spinning (hence "spinnable"). See ReentrantCLHLock that provides a reentrant version.

Memory ordering

NonreentrantCLHLock has the same semantics as ReentrantCLHLock provided that each hand-off of the lock between tasks (if any) establishes a happened-before edge.

Supported operations

acquire(lock::NonreentrantCLHLock; [nspins::Integer]) (lock)

try_race_acquire(lock::NonreentrantCLHLock) (trylock): Not very efficient but lock-free. Fail with AcquiredByWriterError.

release(lock::NonreentrantCLHLock) (unlock)

Locks.ReentrantBackoffSpinLock — Type

ReentrantBackoffSpinLock

A reentrant exponential backoff spin lock.

See also NonreentrantBackoffSpinLock that provides a non-reentrant version.

Extended help

Memory ordering

A release invocation on a lock establishes happened-before edges to subsequent invocations of acquire and try_race_acquire that returns an Ok on the same lock.

Supported operations

ReentrantBackoffSpinLock(; [mindelay], [maxdelay]) -> lock: Create a lock. mindelay (default: 1) specifies the number of spinloop called in the initial backoff. mindelay (default: 1000) specifies the maximum backoff.

acquire(lock::ReentrantBackoffSpinLock; [mindelay], [maxdelay]) (lock): Acquire the lock. Keyword arguments mindelay and maxdelay can be passed to override the values specified by the constructor.

try_race_acquire(lock::ReentrantBackoffSpinLock) (trylock):

release(lock::ReentrantBackoffSpinLock) (unlock)

Locks.NonreentrantBackoffSpinLock — Type

NonreentrantBackoffSpinLock

A non-reentrant exponential backoff spin lock.

See also ReentrantBackoffSpinLock that provides a reentrant version.

Extended help

NonreentrantBackoffSpinLock performs better than Base.Threads.SpinLock with high contention. NonreentrantCLHLock is better than NonreentrantBackoffSpinLock with high contention with many worker threads (20 to 80; it depends on the machine).

Since NonreentrantBackoffSpinLock does not have a fallback "cooperative" waiting mechanism, NonreentrantCLHLock is in general recommended.

Memory ordering

NonreentrantBackoffSpinLock has the same semantics as ReentrantBackoffSpinLock provided that each hand-off of the lock between tasks (if any) establishes a happened-before edge.

Supported operations

NonreentrantBackoffSpinLock(; [mindelay], [maxdelay]) -> lock: Create a lock. mindelay (default: 1) specifies the number of spinloop called in the initial backoff. mindelay (default: 1000) specifies the maximum backoff.

acquire(lock::NonreentrantBackoffSpinLock; [mindelay], [maxdelay]) (lock): Acquire the lock. Keyword arguments mindelay and maxdelay can be passed to override the values specified by the constructor.

try_race_acquire(lock::NonreentrantBackoffSpinLock) (trylock):

release(lock::NonreentrantBackoffSpinLock) (unlock)

Locks.TaskObliviousLock — Type

TaskObliviousLock

A lock that can be released in a task that did not acquire the lock. It does not support reentrancy.

Extended help

Examples

julia> using Locks

julia> lock = TaskObliviousLock();

julia> acquire(lock);

julia> wait(Threads.@spawn release(lock));  # completes

Supported operations

acquire(lock::TaskObliviousLock) (lock)

release(lock::TaskObliviousLock) (unlock)

Implementation detail

TaskObliviousLock is an alias to unspecified implementation of lock. Currently, it is:

julia> TaskObliviousLock === NonreentrantCLHLock
true

Locks.acquire — Function

acquire(lock)

Acquire a lock. It is equivalent to Base.lock(lock) but it may support additional keyword arguments.

See also release and try_race_acquire.

Extended help

Examples

julia> using Locks

julia> lock = ReentrantCLHLock();

julia> acquire(lock);

julia> release(lock);

On naming

Locks.jl uses acquire/release instead of lock/unlock so that:

Variable lock can be used.

Make it clear that Locks.try_race_acquire(lock) -> result::Union{Ok,Err} and Base.trylock(lock) -> locked::Bool have different return types. In particular, try_race_acquire can report the reason why certain attempt have failed.

Locks.release — Function

release(lock)

Release a lock. It is equivalent to Base.unlock(lock).

See also acquire and try_race_acquire.

Locks.try_race_acquire — Function

try_race_acquire(lock) -> Ok(nothing) or Err(reason)

Try to acquire lock and return Ok(nothing) on success. Return an Err wrapping a value explaining a reason of failure.

See the documentation of typeof(lock) for possible error types.

Examples

julia> using Locks

julia> lock = NonreentrantCLHLock();

julia> try_race_acquire(lock)
Try.Ok: nothing

julia> try_race_acquire(lock)
Try.Err: TooManyTries(0, 0)

Locks.race_acquire — Function

race_acquire(lock) -> isacquired::Bool

Try to acquire lock and return true on success and false on failure.

See also try_race_acquire.

Examples

julia> using Locks

julia> lock = NonreentrantCLHLock();

julia> race_acquire(lock)
true

julia> race_acquire(lock)
false

Locks.acquire_then — Function

acquire_then(f, lock; acquire_options...) -> y

Execute a thunk f in a critical section protected by lock and return the value y returned from f. Keyword arguments are passed to acquire.

Extended help

Examples

julia> using Locks

julia> lock = ReentrantCLHLock();

julia> acquire_then(lock; nspins = 10) do
           123 + 456
       end
579